Extract mpu and staging part logic into mpu.py
